# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "KubeFin Cost Analyzer API.",
        "title": "KubeFin API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/costs/clusters/{cluster_id}/namespace": {
            "get": {
                "description": "Get specific cluster namespace costs with time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Get specific cluster namespace costs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The start time to query",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The end time to query",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The step seconds of the data to return",
                        "name": "stepSeconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCostList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/costs/clusters/{cluster_id}/resource": {
            "get": {
                "description": "Get detailed information on cluster resource costs, including CPU, memory, and different billing modes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Get detailed information on cluster resource costs, including CPU, memory, and different billing modes.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The start time to query",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The end time to query",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The step seconds of the data to return",
                        "name": "stepSeconds",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourceCostList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/costs/clusters/{cluster_id}/summary": {
            "get": {
                "description": "Get specific cluster costs summary in current two month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Get specific cluster costs summary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterCostsSummary"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/costs/clusters/{cluster_id}/workload": {
            "get": {
                "description": "Get specific cluster workloads costs with time range",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Get specific cluster workloads costs",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Id",
                        "name": "cluster_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The start time to query",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The end time to query",
                        "name": "endTime",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "The step seconds of the data to return",
                        "name": "stepSeconds",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "The aggregated way to show workload costs",
                        "name": "aggregateBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCostList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
        "/costs/summary": {
            "get": {
                "description": "Get all clusters costs summary in current two month",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Costs"
                ],
                "summary": "Get all clusters costs summary",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterCostsSummaryList"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.StatusError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_kubefin_kubefin_pkg_api.ClusterCostsSummary": {
            "type": "object",
            "properties": {
                "ClusterAvgHourlyCoreCost": {
                    "description": "ClusterAvgDailyCost means average core/hour costs in current month",
                    "type": "number"
                },
                "cloudProvider": {
                    "type": "string"
                },
                "clusterActiveTime": {
                    "description": "ClusterActiveTime shows the cluster active time in seconds",
                    "type": "number"
                },
                "clusterAvgDailyCost": {
                    "description": "ClusterAvgDailyCost means average daily costs in current month",
                    "type": "number"
                },
                "clusterConnectionSate": {
                    "description": "ClusterConnectionSate can be running/connect_failed",
                    "type": "string"
                },
                "clusterId": {
                    "type": "string"
                },
                "clusterMonthCostCurrent": {
                    "description": "ClusterMonthCostCurrent means the total cost in current month",
                    "type": "number"
                },
                "clusterMonthEstimateCost": {
                    "description": "ClusterMonthEstimateCost means the estimating cost with previous 7 days costs",
                    "type": "number"
                },
                "clusterName": {
                    "type": "string"
                },
                "clusterRegion": {
                    "type": "string"
                },
                "connectionTime": {
                    "description": "ConnectionTime shows the time the cluster connected",
                    "type": "integer"
                },
                "lastActiveTime": {
                    "type": "integer"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterCostsSummaryList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterCostsSummary"
                    }
                }
            }
        },
...
        "github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCost": {
            "type": "object",
            "properties": {
                "costList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCostDetail"
                    }
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCostDetail": {
            "type": "object",
            "properties": {
                "cpuCoreUsage": {
                    "type": "number"
                },
                "cpuRequest": {
                    "type": "number"
                },
                "podCount": {
                    "description": "PodCount means the average pod count in this period",
                    "type": "number"
                },
                "ramGBRequest": {
                    "type": "number"
                },
                "ramGBUsage": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "totalCost": {
                    "type": "number"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCostList": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterNamespaceCost"
                    }
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterResourceCost": {
            "type": "object",
            "properties": {
                "costFallbackBillingMode": {
                    "description": "TODO: implement this type",
                    "type": "number"
                },
                "costOnDemandBillingMode": {
                    "type": "number"
                },
                "costPeriodBillingMode": {
                    "type": "number"
                },
                "costSpotBillingMode": {
                    "type": "number"
                },
                "cpuCoreCount": {
                    "description": "CPUCoreCount means the average core hour count in this period",
                    "type": "number"
                },
                "cpuCoreUsage": {
                    "description": "CPUCoreUsage means the average core hour usage in this period",
                    "type": "number"
                },
                "cpuCost": {
                    "type": "number"
                },
                "ramCost": {
                    "type": "number"
                },
                "ramGBCount": {
                    "description": "RAMGBCount means the average ram hour count in this period",
                    "type": "number"
                },
                "ramGBUsage": {
                    "description": "RAMGBUsage means the average ram hour usage in this period",
                    "type": "number"
                },
                "timestamp": {
                    "description": "Timestamp is in unix timestamp format, you can transform it to any you want",
                    "type": "integer"
                },
                "totalCost": {
                    "type": "number"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterResourceCostList": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterResourceCost"
                    }
                }
            }
        },
...
        "github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCost": {
            "type": "object",
            "properties": {
                "costList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCostDetail"
                    }
                },
                "namespace": {
                    "type": "string"
                },
                "workloadName": {
                    "type": "string"
                },
                "workloadType": {
                    "description": "WorkloadType could be pod/daemonset/statefulset/deployment",
                    "type": "string"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCostDetail": {
            "type": "object",
            "properties": {
                "cpuCoreRequest": {
                    "type": "number"
                },
                "cpuCoreUsage": {
                    "type": "number"
                },
                "podCount": {
                    "description": "PodCount means the average pod count in this period",
                    "type": "number"
                },
                "ramGBRequest": {
                    "type": "number"
                },
                "ramGBUsage": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "totalCost": {
                    "type": "number"
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCostList": {
            "type": "object",
            "properties": {
                "clusterId": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_kubefin_kubefin_pkg_api.ClusterWorkloadCost"
                    }
                }
            }
        },
        "github_com_kubefin_kubefin_pkg_api.StatusError": {
            "type": "object",
            "properties": {
                "apiVersion": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "kind": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}